html/semantics/forms/the-form-element/form-nameditem.html

<!doctype html>
<meta charset=utf-8>
<title>Form named getter</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<form>
<input type=button name=button>
<input type=radio name=radio value=x>
<input type=radio name=radio value=y>
<input type=radio name=radio value=z>
</form>
<form>
<!-- EventTarget -->
<input type=radio name=addEventListener>
<input type=radio name=removeEventListener>
<input type=radio name=dispatchEvent>
<!-- Node -->
<input type=radio name=nodeType>
<input type=radio name=nodeName>
<input type=radio name=ownerDocument>
<!-- Element -->
<input type=radio name=namespaceURI>
<input type=radio name=prefix>
<input type=radio name=localName>
<!-- HTMLElement -->
<input type=radio name=title>
<input type=radio name=lang>
<input type=radio name=dir>
<!-- HTMLFormElement -->
<input type=radio name=acceptCharset>
<input type=radio name=action>
<input type=radio name=autocomplete>
<input type=radio name=enctype>
<input type=radio name=encoding>
<input type=radio name=method>
<input type=radio name=name>
<input type=radio name=noValidate>
<input type=radio name=target>
<input type=radio name=elements>
<input type=radio name=length>
<input type=radio name=submit>
<input type=radio name=reset>
<input type=radio name=checkValidity>
<script>
test(function() {
  var form = document.getElementsByTagName("form")[0]
  assert_equals(form.item, undefined)
  assert_false("item" in form)
}, "Forms should not have an item method")
test(function() {
  var form = document.getElementsByTagName("form")[0]
  assert_equals(form.namedItem, undefined)
  assert_false("namedItem" in form)
}, "Forms should not have a namedItem method")
test(function() {
  var form = document.getElementsByTagName("form")[0]
  var button = document.getElementsByTagName("input")[0]
  assert_equals(button.type, "button")
  assert_equals(form.button, button)
  assert_equals(form.button.length, undefined)
}, "Name for a single element should work")
test(function() {
  var form = document.getElementsByTagName("form")[0]
  assert_equals(form.radio.item(-1), null)
  assert_array_equals([0, 1, 2].map(function(i) {
    return form.radio.item(i).value
  }), ["x", "y", "z"])
  assert_equals(form.radio.item(3), null)
}, "Calling item() on the NodeList returned from the named getter should work")
test(function() {
  var form = document.getElementsByTagName("form")[0]
  assert_equals(form.radio.length, 3)
  assert_equals(form.radio[-1], undefined)
  assert_array_equals([0, 1, 2].map(function(i) {
    return form.radio[i].value
  }), ["x", "y", "z"])
  assert_equals(form.radio[3], undefined)
}, "Indexed getter on the NodeList returned from the named getter should work")
test(function() {
  var form = document.getElementsByTagName("form")[0]
  var indices = [-1, 0, 1, 2, 3]
  indices.forEach(function(i) {
    assert_throws(new TypeError(), function() {
      form.radio(i)
    })
  })
}, "Invoking a legacycaller on the NodeList returned from the named getter " +
   "should not work")
test(function() {
  var names = [
    // EventTarget
    "addEventListener", "removeEventListener", "dispatchEvent",
    // Node
    "nodeType", "nodeName", "ownerDocument",
    // Element
    "namespaceURI", "prefix", "localName",
    // HTMLElement
    "title", "lang", "dir",
    // HTMLFormElement
    "acceptCharset", "action", "autocomplete", "enctype", "encoding", "method",
    "name", "noValidate", "target", "elements", "length", "submit", "reset",
    "checkValidity"
  ]
  var form = document.getElementsByTagName("form")[1]
  names.forEach(function(name, i) {
    assert_equals(form[name], form.children[i])
  })
}, "Named elements should override builtins")
</script>
